--- a/spec/03-plugins/25-oauth2/02-api_spec.lua
+++ b/spec/03-plugins/25-oauth2/02-api_spec.lua
@@ -1,16 +1,16 @@
 local cjson   = require "cjson"
 local helpers = require "spec.helpers"
+local admin_api = require "spec.fixtures.admin_api"
 
 
 for _, strategy in helpers.each_strategy() do
   describe("Plugin: oauth (API) [#" .. strategy .. "]", function()
-    local consumer
-    local service
     local admin_client
-    local bp, db
+    local db
 
     lazy_setup(function()
-      bp, db = helpers.get_db_utils(strategy, {
+      local _
+      _, db = helpers.get_db_utils(strategy, {
         "routes",
         "services",
         "consumers",
@@ -37,14 +37,18 @@ for _, strategy in helpers.each_strategy() do
     end)
 
     describe("/consumers/:consumer/oauth2/", function()
+      local consumer
+      local service
+
       lazy_setup(function()
-        service = bp.services:insert({ host = "oauth2_token.com" })
-        consumer = bp.consumers:insert({ username = "bob" })
-        bp.consumers:insert({ username = "sally" })
+        service = admin_api.services:insert({ host = "oauth2_token.com" })
+        consumer = admin_api.consumers:insert({ username = "bob" })
+        admin_api.consumers:insert({ username = "sally" })
       end)
 
-      after_each(function()
-        assert(db:truncate("oauth2_credentials"))
+      lazy_teardown(function()
+        admin_api.consumers:remove(consumer)
+        admin_api.services:remove(service)
       end)
 
       describe("POST", function()
@@ -278,9 +282,13 @@ for _, strategy in helpers.each_strategy() do
       end)
 
       describe("GET", function()
+        local consumer = admin_api.consumers:insert({
+          username = "get_test",
+        })
+        local credentials = {}
         lazy_setup(function()
           for i = 1, 3 do
-            bp.oauth2_credentials:insert {
+            credentials[i] = admin_api.oauth2_credentials:insert {
               name          = "app" .. i,
               redirect_uris = { helpers.mock_upstream_ssl_url },
               consumer      = { id = consumer.id },
@@ -288,12 +296,14 @@ for _, strategy in helpers.each_strategy() do
           end
         end)
         lazy_teardown(function()
-          assert(db:truncate("oauth2_credentials"))
+          for _, credential in ipairs(credentials) do
+            admin_api.oauth2_credentials:remove(credential)
+          end
         end)
         it("retrieves the first page", function()
           local res = assert(admin_client:send {
             method  = "GET",
-            path    = "/consumers/bob/oauth2"
+            path    = "/consumers/get_test/oauth2"
           })
           local body = assert.res_status(200, res)
           local json = cjson.decode(body)
@@ -305,24 +315,31 @@ for _, strategy in helpers.each_strategy() do
 
     describe("/consumers/:consumer/oauth2/:id", function()
       local credential
+      local consumer
+      local service
 
       lazy_setup(function()
-        assert(db:truncate("routes"))
-        assert(db:truncate("services"))
-        assert(db:truncate("consumers"))
-        service = bp.services:insert({ host = "oauth2_token.com" })
-        consumer = bp.consumers:insert({ username = "bob" })
+        service = admin_api.services:insert({ host = "oauth2_token.com" })
+        consumer = admin_api.consumers:insert({ username = "bob" })
+      end)
+
+      lazy_teardown(function()
+        admin_api.consumers:remove(consumer)
+        admin_api.services:remove(service)
       end)
 
       before_each(function()
-        assert(db:truncate("oauth2_credentials"))
-
-        credential = bp.oauth2_credentials:insert {
+        credential = admin_api.oauth2_credentials:insert {
           name          = "test app",
           redirect_uris = { helpers.mock_upstream_ssl_url },
           consumer      = { id = consumer.id },
         }
       end)
+
+      after_each(function()
+        admin_api.oauth2_credentials:remove(credential)
+      end)
+
       describe("GET", function()
         it("retrieves oauth2 credential by id", function()
           local res = assert(admin_client:send {
@@ -343,9 +360,12 @@ for _, strategy in helpers.each_strategy() do
           assert.equal(credential.id, json.id)
         end)
         it("retrieves credential by id only if the credential belongs to the specified consumer", function()
-          bp.consumers:insert {
+          local alice = admin_api.consumers:insert {
             username = "alice"
           }
+          finally(function()
+            admin_api.consumers:remove(alice)
+          end)
 
           local res = assert(admin_client:send {
             method  = "GET",
@@ -457,8 +477,17 @@ for _, strategy in helpers.each_strategy() do
 
     describe("/oauth2", function()
       describe("POST", function()
+        local consumer
+        local service
+
         lazy_setup(function()
-          db:truncate("oauth2_credentials")
+          service = admin_api.services:insert({ host = "oauth2_token.com" })
+          consumer = admin_api.consumers:insert({ username = "bob" })
+        end)
+
+        lazy_teardown(function()
+          admin_api.consumers:remove(consumer)
+          admin_api.services:remove(service)
         end)
 
         it("does not create oauth2 credential when missing consumer", function()
@@ -502,8 +531,17 @@ for _, strategy in helpers.each_strategy() do
 
     describe("/oauth2/:client_id_or_id", function()
       describe("PUT", function()
+        local consumer
+        local service
+
         lazy_setup(function()
-          db:truncate("oauth2_credentials")
+          service = admin_api.services:insert({ host = "oauth2_token.com" })
+          consumer = admin_api.consumers:insert({ username = "bob" })
+        end)
+
+        lazy_teardown(function()
+          admin_api.consumers:remove(consumer)
+          admin_api.services:remove(service)
         end)
 
         it("does not create oauth2 credential when missing consumer", function()
@@ -547,19 +585,26 @@ for _, strategy in helpers.each_strategy() do
     end)
 
     describe("/oauth2_tokens/", function()
-      local oauth2_credential
-      lazy_setup(function()
-        oauth2_credential = bp.oauth2_credentials:insert {
-          name          = "Test APP",
-          redirect_uris = { helpers.mock_upstream_ssl_url },
-          consumer      = { id = consumer.id },
-        }
-      end)
-      after_each(function()
-        assert(db:truncate("oauth2_tokens"))
-      end)
-
       describe("POST", function()
+        local oauth2_credential
+        local consumer
+        local service
+
+        lazy_setup(function()
+          service = admin_api.services:insert({ host = "oauth2_token.com" })
+          consumer = admin_api.consumers:insert({ username = "bob" })
+          oauth2_credential = admin_api.oauth2_credentials:insert {
+            name          = "Test APP",
+            redirect_uris = { helpers.mock_upstream_ssl_url },
+            consumer      = { id = consumer.id },
+          }
+        end)
+
+        lazy_teardown(function()
+          admin_api.consumers:remove(consumer)
+          admin_api.services:remove(service)
+        end)
+
         it("creates a oauth2 token", function()
           local res = assert(admin_client:send {
             method  = "POST",
@@ -602,19 +647,34 @@ for _, strategy in helpers.each_strategy() do
       end)
 
       describe("GET", function()
+        local oauth2_credential
+        local consumer
+        local service
+
         lazy_setup(function()
-          for _ = 1, 3 do
-            bp.oauth2_tokens:insert {
+          service = admin_api.services:insert({ host = "oauth2_token.com" })
+          consumer = admin_api.consumers:insert({ username = "bob" })
+          oauth2_credential = admin_api.oauth2_credentials:insert {
+            name          = "Test APP",
+            redirect_uris = { helpers.mock_upstream_ssl_url },
+            consumer      = { id = consumer.id },
+          }
+        end)
+
+        lazy_teardown(function()
+          admin_api.consumers:remove(consumer)
+          admin_api.services:remove(service)
+        end)
+
+        it("retrieves the first page", function()
+          for i = 1, 3 do
+            admin_api.oauth2_tokens:insert {
               credential = { id = oauth2_credential.id },
               service    = { id = service.id },
               expires_in = 10
             }
           end
-        end)
-        lazy_teardown(function()
-          assert(db:truncate("oauth2_tokens"))
-        end)
-        it("retrieves the first page", function()
+
           local res = assert(admin_client:send {
             method  = "GET",
             path    = "/oauth2_tokens"
@@ -627,14 +687,35 @@ for _, strategy in helpers.each_strategy() do
       end)
 
       describe("/oauth2_tokens/:id", function()
+        local oauth2_credential
+        local consumer
+        local service
+
+        lazy_setup(function()
+          service = admin_api.services:insert({ host = "oauth2_token.com" })
+          consumer = admin_api.consumers:insert({ username = "bob" })
+          oauth2_credential = admin_api.oauth2_credentials:insert {
+            name          = "Test APP",
+            redirect_uris = { helpers.mock_upstream_ssl_url },
+            consumer      = { id = consumer.id },
+          }
+        end)
+
+        lazy_teardown(function()
+          admin_api.consumers:remove(consumer)
+          admin_api.services:remove(service)
+        end)
+
         local token
         before_each(function()
-          assert(db:truncate("oauth2_tokens"))
           token = db.oauth2_tokens:insert {
             credential = { id = oauth2_credential.id },
             service    = { id = service.id },
             expires_in = 10
           }
+        end)
+        after_each(function()
+          admin_api.oauth2_tokens:remove(token)
         end)
 
         describe("GET", function()