Fix test flakiness issue discussed in https://github.com/Kong/kong/pull/5354 (and also make them faster)

As a bonus, speed up the Admin API tests for Certificates too: on my machine, Cassandra tests for it went from 34s to 2.8s (woohoo!)

Aaand while I was waiting for CI to run on other PRs, I sped up the OAuth2 tests too: on my machine, Cassandra tests went from 52s to 27s :+1: